requirements.txt
pymongo>=4.8.0
Faker>=28.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0